Rename endpoints module to endpoint
